+++ spec/lib/kitchen/terraform/command/get_spec.rb
@@ -16,7 +16,7 @@
 
 require "kitchen"
 require "kitchen/terraform/command/get"
-require "kitchen/terraform/shell_out_nu"
+require "kitchen/terraform/shell_out"
 
 ::RSpec.describe ::Kitchen::Terraform::Command::Get do
   describe ".run" do
@@ -37,7 +37,7 @@ require "kitchen/terraform/shell_out_nu"
     end
 
     before do
-      allow(::Kitchen::Terraform::ShellOutNu).to receive(:run_command).with(
+      allow(::Kitchen::Terraform::ShellOut).to receive(:run_command).with(
         "terraform get -update",
         cwd: directory,
         environment: kind_of(::Hash),
