+++ spec/lib/kitchen/terraform/command/version_spec.rb
@@ -16,7 +16,7 @@
 
 require "kitchen"
 require "kitchen/terraform/command/version"
-require "kitchen/terraform/shell_out_nu"
+require "kitchen/terraform/shell_out"
 
 ::RSpec.describe ::Kitchen::Terraform::Command::Version do
   describe ".run" do
@@ -25,7 +25,7 @@ require "kitchen/terraform/shell_out_nu"
     end
 
     before do
-      allow(::Kitchen::Terraform::ShellOutNu).to receive(:run).with(command: "terraform version").and_yield(
+      allow(::Kitchen::Terraform::ShellOut).to receive(:run).with(command: "terraform version").and_yield(
         output: output,
       )
     end
